Input folder is now flat
