+++ src/Input/Input.styles.tsx
@@ -1,0 +1,14 @@
+export const styles = {
+  input: {
+    height: '36px',
+    fontSize: '15px',
+    outline: 'none',
+    borderRadius: '24px',
+    border: '1px solid #d9d9d9',
+    padding: '0px 12px',
+    // boxSizing: 'border-box',
+  },
+  focusInput: {
+    border: '1px solid #1890ff',
+  },
+};
